commandcenter/agents/memory_pruner/requirements.txt
psycopg2-binary==2.9.9
qdrant-client==1.7.0
redis==5.0.1
schedule==1.2.0
